04112021-538/requirements.txt
annotated-types==0.7.0
anthropic==0.38.0
anyio==4.6.2.post1
attrs==24.2.0
backports.tarfile==1.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.8.30
chardet==4.0.0
charset-normalizer==3.4.0
dataclasses==0.6
deprecation==2.1.0
distro==1.9.0
exceptiongroup==1.2.2
filelock==3.16.1
fsspec==2024.10.0
gotrue==0.2.0
h11==0.14.0
httpcore==0.12.3
httpx==0.16.1
huggingface-hub==0.26.2
idna==2.10
importlib_metadata==8.5.0
iniconfig==2.0.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jiter==0.7.0
keyring==25.5.0
more-itertools==10.5.0
openai==1.53.0
outcome==1.3.0.post0
packaging==24.1
pluggy==1.5.0
postgrest-py==0.4.0
py==1.11.0
pydantic==2.9.2
pydantic_core==2.23.4
PySocks==1.7.1
pytest==6.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
realtime-py==0.1.3
requests==2.25.1
rfc3986==1.5.0
selenium==4.26.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
supabase-py==0.0.2
tokenizers==0.20.1
toml==0.10.2
tqdm==4.66.6
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==1.26.20
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==9.1
wsproto==1.2.0
zipp==3.20.2
